Add scenatio updating a project
